refactor: split runtime governance and mode execution
